backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==1.10.7
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
websockets==11.0.1
scikit-learn==1.2.2
pandas==2.0.0
numpy==1.24.2
pytest==7.3.1
httpx==0.24.0
motor==3.1.1
email-validator==2.0.0
Pillow==9.5.0
python-slugify==8.0.1
stripe==5.4.0
redis==4.5.4
celery==5.2.7
boto3==1.26.123
requests==2.28.2
aiofiles==23.1.0
jinja2==3.1.2
itsdangerous==2.1.2
pyjwt==2.6.0
matplotlib==3.7.1
seaborn==0.12.2
statsmodels==0.13.5
openpyxl==3.1.2
xlrd==2.0.1
gunicorn==20.1.0